refactor: Move request utils to fixtures
