+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/service/CandidateTestSetup.java
@@ -1,8 +1,8 @@
 package no.sikt.nva.nvi.common.service;
 
+import static no.sikt.nva.nvi.common.UpsertRequestBuilder.randomUpsertRequestBuilder;
 import static no.sikt.nva.nvi.common.db.PeriodRepositoryFixtures.periodRepositoryReturningOpenedPeriod;
 import static no.sikt.nva.nvi.test.TestUtils.randomBigDecimal;
-import static no.sikt.nva.nvi.test.UpsertRequestBuilder.randomUpsertRequestBuilder;
 import static no.unit.nva.testutils.RandomDataGenerator.randomUri;
 import static org.mockito.Mockito.mock;
